several improvements on vochaintest and client package
Reuse previous census if the electionSize is the same.
Fix a bug with the vote splitting.
Rename some client functions to Test***.
Do not import census manually on gateways but instead wait until they
got it automatically.
Use cancelProcessTX instead of waiting for process to end.
Make max time dynamic depending on the electionSize.
Some other minor improveents.

Example of usage on the development testnet:

 ORACLE=0x89.... (ask me for the priv key)
 EXTRA=wss://gw4.dev.vocdoni.net/dvote,wss://gw1.dev.vocdoni.net/dvote,wss://gw3.dev.vocdoni.net/dvote
 go run cmd/vochaintest/vochaintest.go --oracleKey=$ORACLE --gwHost=wss://gw1.dev.vocdoni.net/dvote --gwExtra=$EXTRA --parallelCons=3 --electionSize=2000 --doubleVote=False

Signed-off-by: p4u <pau@dabax.net>