net: support proxying json-rpc via a socket
This is useful when W3External points at a socket file, e.g. from
openethereum.

The implementation is similar to the one that points to an HTTP
endpoint; we create a new connection for each request. In the future, we
could optimize this by reusing a small number of connections.

Fixes #293.