net: keep HTTP connection open until we write a response
Fixes an issue where we would fail to send any response. The reason is
that, when the handler func returns, the server will return an empty
response to the client, such as a 200 without a body.

So, when we would try to write our response later, we would get all
sorts of weird errors.

To fix this, use a channel to signal when we're done writing our
response. Use a select with chi's 30s timeout, so that we don't
potentially leak goroutines forever if we fail to respond.