ipfssync: use a timeout when listing pins
We were using context.TODO for listing all pins; this can be an
expensive operation, and if the network or any other nodes are being
unreliable, it could potentially hang for hours. A gateway crashed with
an OOM panic, and thousands of "list pins" ipfs goroutines were hung in
selects for 7000+ minutes.

The ipfs component already has a timeout, which defaults to 600s, or ten
minutes. That should be plenty for any non-trivial operation or large
amount of pins, but it should also be small enough to prevent thousands
of goroutines piling up over many hours.

Fixes #300, hopefully.