docker: use buildkit, cache mod and build with --mount
The biggest advantage is that the dependency download and package build
caches are now shared transparently between builds, without being an
all-or-nothing deal.

Before, changing a single line in go.mod (such as adding a dependency)
would re-trigger the 'go mod download' line, and thus we would have to
download all the dependencies again.

Similarly, changing a single line of Go would trigger the 'go build'
line, and thus would require rebuilding everything from scratch.

Now, those two caches are mounted, so both of those issues are gone.
Adding a single dependency or changing a single Go file should now make
'docker build' take under ten seconds, not many minutes.

Beware that this requires 'docker buildx', AKA buildkit, which has been
available with Docker since version 18.09, released in September 2019 -
almost a year ago. I just checked that CI is using the latest stable
version for "docker in docker", which is 19.03.12 as of last week.

Buildkit is not enabled by default, but can be easily enabled with
either of the methods below, as per
https://docs.docker.com/develop/develop-images/build_enhancements/:

	# method 1; only affects the current shell session
	$ export DOCKER_BUILDKIT=1

	# method 2; affects the Docker daemon permanently
	$ cat /etc/docker/daemon.json
	{ "features": { "buildkit": true } }

On top of the change above, docker-compose needs to be told to use the
docker build CLI tool, to make use of buildkit as well:

	$ COMPOSE_DOCKER_CLI_BUILD=1 docker-compose build