add ResolveEntityMetadataURL function, make ensResolve public
ResolveEntityMetadata URL fetches the text on the ENS resolver
that contains the IPFS URL where the entity metadata is stored.

Make ensResolve funcion public for better location, is not part
of the ethereum service but a part of the chain pkg itself.
Also needs to be available for other packages (service).