use []byte instead of string for processID and entityID
This change afects a lot of code, but it makes a lot of sense.
There are still a lot of types that should be moved from string to its
real type, but that can be done slowly.

This changes are not backwards compatible, so the Vochain consensus
state will eventually break. So development genesis is updated already
on this commit.

Signed-off-by: p4u <pau@dabax.net>