types: stop importing tendermint/types
This means that our types package is far lighter and more portable,
since the tendermint package indirectly imports heavy packages such as
grpc or crypto implementations.

This also means that all the types which we encode as part of our app
state are now controlled by us, which is something we should have done
since the beginning anyway. Before this change, updating the tendermint
version could simply break the state.

Two tricky changes needed to be done. First, we had to drop the Equal
method on the PubKey interface; see the added comment. Second, we had to
tweak the amino code a little bit, to also register our copied interface.