change the websocket client mechanism and package
Drop wsPool and use instead recws, a gorilla websocket client wrapper
that handles automatically the reconnection

It uses github.com/recws-org/recws but there is a bug on the
reconnection handling. This bug is fixed on my temporary fork
github.com/p4u/recws, the PR has been sent but while its not merged we
have to use mine.

Signed-off-by: p4u <pau@dabax.net>