cmd/dvotenode: print own version when starting
This version is statically included into the binary, via a linker flag
and "git describe". It is automatically included with the Docker builds.

Because Docker now needs the git folder, it's removed from dockerignore.

Now, starting a clean dvote node starts with something like:

	vocdoni dvote version v0.6.0-43-gc851e43-dirty
	no signing key, generating one...
	<date> INFO dvotenode/dvotenode.go:322 logger construction succeeded at level info and output stdout
	<date> WARN dvotenode/dvotenode.go:334 non Critical error loading config: creating new config file in /root/.dvote
	<date> INFO dvotenode/dvotenode.go:368 starting vocdoni dvote node version v0.6.0-43-gc851e43-dirty in gateway mode

The reason the version is printed twice is explained in the code
comment. We include the version in a new "internal" package, so that
other main packages can also use it without exposing it to third party
modules.