enable IndexAllKeys Vochain
This is required in order to get a bulk of txs given block height.
Until now the only index used was the tx.hash because this feature
was not required.
Also we usually make app state queries via IAVL and API calls
to avoid using the ABCI Query method. But given this not use
the app state we can obtain a list of txs by height using the
tx_seach Tendermint RPC method.

So let's set IndexAllKeys to true and enable the tx indexing
both by hash and height.