all: start using testing.TB.TempDir
This method was introduced in Go 1.15, which we've been using for three
months, so we can safely start using it.

We had a similar implementation copy-pasted in a handful of places, so
we get to delete a significant amount of code too.

While at it, testing.B should be called "b", not "t".