router: use timeouts for all IPFS operations
Four IPFS operations were still using a context.TODO, meaning that there
was no timeout set for the operation whatsoever.

We were seeing IPFS goroutines leaking for many hours at a time:

	goroutine 1227617643 [select, 4959 minutes]:
	github.com/ipfs/go-ipfs/core/coreapi.(*PinAPI).pinLsAll.func1(0xc0c1b18000, 0x841, 0x841, 0x2d2b87d, 0x9, 0x841, 0x0)
		/go/pkg/mod/github.com/ipfs/go-ipfs@v0.6.0/core/coreapi/pin.go:230 +0x325
	github.com/ipfs/go-ipfs/core/coreapi.(*PinAPI).pinLsAll.func3(0xc025794cc0, 0x2d20daa, 0x3, 0xc001c12a00, 0x31fda20, 0xc000056058, 0xc0081aa210, 0xc00f0b8540)
		/go/pkg/mod/github.com/ipfs/go-ipfs@v0.6.0/core/coreapi/pin.go:258 +0x1d8
	created by github.com/ipfs/go-ipfs/core/coreapi.(*PinAPI).pinLsAll
		/go/pkg/mod/github.com/ipfs/go-ipfs@v0.6.0/core/coreapi/pin.go:249 +0x1a5

This was most likely caused by the r.storage.ListPins call without a
timeout, which could get stuck for one reason or another (such as a
disconnected peer), and then we'd leak the goroutine forever.

Instead, use a standard timeout. Ten seconds was already being used
elsewhere in the same file and it seems reasonable, so use that. Give it
a better name, too.

Fixes #300 again, hopefully for good this time.