net: don't panic if an http request gets cancelled
For example, if handling a request takes a while and the router times
out first, then the HTTP handler would exit before the Send call
happened. Then we would get a panic:

	panic: Header called after Handler finished

The fix is rather simple; since every router callback must call Send
before it's finished, we don't exit the handler func until that has
happened.

In our original case, Send would then be called once the request has
been cancelled. It would not panic as per our fix above, but it would
still log unhelpful errors, since writing a response to a closed client
connection will fail. For this reason, if Send is called and the
connection was closed, do nothing. But still close the "sent" channel,
so that the handler can exit.

Fixes #331.