add MetaRequest.String for easier log prints
It does something similar to %v, but it omits zero fields and prints
[]byte as hexadecimal.

With a few test cases too, because this is too easy to write in the
wrong way or break. In the future we could make this code reusable for
other types with lots of fields which are often empty.