data: finally fix the elusive ipfs goroutine leak
We thought we had fixed the goroutine leak below after adding contexts
with timeouts everywhere, but we were still getting less severe leaking
over long periods of time:

	goroutine 431409 [chan send, 163 minutes]:
	github.com/ipfs/go-ipfs/core/coreapi.(*PinAPI).pinLsAll.func3(0xc04e58ad20, 0x307836f, 0x3, 0xc0030b4c80, 0x3675620, 0xc04ccf5b00, 0xc04e5b8060, 0xc043a56fd0)
		/go/pkg/mod/github.com/ipfs/go-ipfs@v0.7.1-0.20201201031203-895b340e08b9/core/coreapi/pin.go:259 +0x10da
	created by github.com/ipfs/go-ipfs/core/coreapi.(*PinAPI).pinLsAll
		/go/pkg/mod/github.com/ipfs/go-ipfs@v0.7.1-0.20201201031203-895b340e08b9/core/coreapi/pin.go:249 +0x1a5

	goroutine 3878139 [chan send, 43 minutes]:
	github.com/ipfs/go-ipfs/core/coreapi.(*PinAPI).pinLsAll.func3(0xc02a8e3020, 0x307836f, 0x3, 0xc0030b4c80, 0x3675620, 0xc04937c900, 0xc0467c6ff0, 0xc047a28c90)
		/go/pkg/mod/github.com/ipfs/go-ipfs@v0.7.1-0.20201201031203-895b340e08b9/core/coreapi/pin.go:259 +0x10da
	created by github.com/ipfs/go-ipfs/core/coreapi.(*PinAPI).pinLsAll
		/go/pkg/mod/github.com/ipfs/go-ipfs@v0.7.1-0.20201201031203-895b340e08b9/core/coreapi/pin.go:249 +0x1a5

It turns out that the fault was ours. The Pin().Ls API returns a channel
for iterating over the entire list of pins without having to keep the
entire list in memory at once. Unfortunately, our two functions to
collect stats did not use the channel properly; they assumed "pins" was
a slice and not a channel, so they simply did len(pins).

This built and ran, so noone noticed, but it meant that dvotenode always
reported that there were 0 pins, via both the logs and Prometheus:

	[ipfs info] peers:145 addresses:6112 pins:0

The reason is that the length of a channel is simply its buffer size.
Since IPFS returns a non-buffered channel, that is always 0. The leak
was caused because the API tried to send all the results over the
channel, but since we never read from the channel, the API got blocked
forever. Thus, that goroutine leaked.

To fix the leak, and make the metrics correct, we must range over the
channel and count the number of pins. Consuming the entire channel makes
the API implementation close the channel and exit its goroutines, so
there is no longer any leak.

Fixes #332.