refactor vochain internal cache
Since Tendermint now exports methods for deleting a transaction from the
mempool, it is possible to remove old Transactions once they are not
valid anymore or a timeout is triggered.

In addition, convert the previours VoteCache to a general purpose Cache
so it can be used by any kind of blockchain transaction.

The current implementation however only caches the vote transaction.
Next commits will add support for the rest of transactions.

Signed-off-by: p4u <pau@dabax.net>