ipfssync: make copies of keys when iterating over the tree
In the Graviton tree, when iterating over keys/values, the []byte
parameters are only valid during the lifetime of the callback func. The
buffers are reused for future iterations, so we must make copies to keep
them around; otherwise, we have races.

It's unclear if the IAVL tree has the same restriction, but the docs say
that the []byte points to the internal data, so I assume it would have a
very similar data race problem.

	WARNING: DATA RACE
	Write at 0x00c002ccc312 by goroutine 39:
	  runtime.slicecopy()
	      /usr/lib/go/src/runtime/slice.go:246 +0x0
	  [...]

	Previous read at 0x00c002ccc312 by goroutine 523:
	  runtime.slicebytetostring()
	      /usr/lib/go/src/runtime/string.go:80 +0x0
	  go.vocdoni.io/dvote/ipfssync.(*IPFSsync).syncPins()
	      /home/mvdan/src/dvote/ipfssync/ipfssync.go:178 +0x36d
	  go.vocdoni.io/dvote/ipfssync.(*IPFSsync).Start.func4()
	      /home/mvdan/src/dvote/ipfssync/ipfssync.go:431 +0x52