types: use HexBytes instead of hex strings in a few API fields
This means the json marshal/unmarshal steps will automatically do the
hex encoding/decoding, and we can simply use the []byte data directly.

Removes over 100 lines of boilerplate thanks to the above. More "hex
string" fields remain in the API structs, but they are left alone for
now to keep this refactor smaller.

Note that this should be an entirely backwards compatible change, since
the encoded JSON messages should look exactly the same. Replacing REST
with gRPC is still planned for the future, but this is a clear win we
can get today. Plus, our actual business logic gets clearer.

While at it, replace tendermint's bytes package with our own
implementation, which is smaller and ensures it won't change behavior.
With a test, too.