ipfssync: fix send pin list when fetch receied
Previous implementation of fetch handler was wrong since the pin list
was only sent if the remote and local hashes are equal. The behavior
must be the opposite (send the list of hashes are different).

Small changes on subpub: try to avoid "stream reset" problems by using a
different context than the one used in discoer(). In addition call
stream.Close() in case of "stream reset" to ensure the connection is
actually removed.

Signed-off-by: p4u <pau@dabax.net>