census: allow using values (in addition to keys)
Until now only keys were allowed to be used as part of the census merkle
tree. This commit introduces also values, which are stored as a leaf of
the census tree. The values do not change the position of the claim into
the merkle tree, its aim is to be used as a numeric weight.

This is the first step to support weighted elections.

WARNING: API breaking change

Signed-off-by: p4u <pau@dabax.net>