mod: update to the final tendermint 0.34.1 version
This is actually a sort of downgrade, since in early December we
upgraded to the latest "master" version instead of using the 0.34
release branch with just the backported fixes we needed.

As such, we adapted to some breaking 0.35 changes, which we're undoing
here. There won't be a 0.35 stable release for multiple months, so
sticking to stable is the sanest option.