fix up a few of my TODOs
The go-ethereum method to stop a node is no longer racy, so use it
again.

A TODO had a link to a gitlab issue, so just remove it as it's not
really necessary.

We also agreed to leave Payload as it is, and clean it up as part of a
v2 protobuf API.

While at it, switch eth_test.go to quicktest, and simplify
NewMockEthereum.