chain: add oracle Ethereum whitelist for event processing
Add a map of whitelisted addresses on the Oracle
in order to check if the tx comming from the source of
truth is one of the whitelisted. If the address
that created the tx on the source of truth is on the
map and set to true the Oracle will process the event. If not
the event will be ignored.

- add EthereumWhiteListAddrs flag for setting
  the valid addresses, by default all are accepted.
- initialize the events service with the EthereumWhiteListAddrs
- check if the address that created the tx on the source of
  truth is on the oracle Ethereum whitelist before creating the vochain tx

The address of the tx creator on the source of truth is obtained
thanks to the logs. Logs include a tx hash and through
this tx hash we can obtain the tx creator.

Careful: This is not working for internal calls
e.g. Call chain: Contract A -> Contract B -> Contract C
     msg.Sender: C(from) == B != A

e.g. EOA A -> Contract B
     msg.Sender: B(from) == A