bump sha256-simd version to work on Go master again
The upcoming Go 1.17 version gained a new check for Go functions written
in assembly. It seems to be unhappy about the older version of
sha256-simd:

	traceback: unexpected SPWRITE function github.com/minio/sha256-simd.blockAvx2
	fatal error: traceback

Bump the version of this indirect dependency. Upstream very recently
removed the code in question, as the pure Go is faster now.

After this change, 'go test ./...' passes again on:

	go version devel +37ca84a9cd Thu Feb 25 09:46:02 2021 +0000 linux/amd64