fix vochain db fatal on startup:
First, when trying to read priv_validator_state.json,
if empty the Tendermint lib executes an os.Exit internally
as the file does not have any bytes to read.
The os.Exit cannot be handled at a lib level.
The applied fix opens or creates the mentioned file
and adds the default values that Tendermint expects to have,
before the Tendermint lib tries to open the file.

Second, the syncronization of the node starts always at height
N-1. There is an edge case where the node is halted before
the first block is processed. Then when the node is started again
any version exists.
If version N-1 fails to load, also version 0, there is
nothing that can be done rather than restarting the entire database,
thus avoiding corruption on state bytes.
A backup of all the vochain data is stored for security reasons.