use badgerhold as KV database replacement on scrutinizer
Using badgerhold instead of plain KV store (badger) brings a new set of
possibilities for querying the Vochain. We can now easily request
information with a SQL-like style of syntax while preserving the
simplicity and low resource consumption of using a KV store.

This is the first step for renaming the current scrutinizer to indexer
(on a more generic purpose) and for decopling the consensus state from
the API querying database.

In addition to the scrutinizer refactor, change the FromId API field to
From (numeric value) since the processes and entities are now listed by
creation time. Also add a new API field named "Namespace" which can be
used for querying on ProcessList for a specific Namespace (0 means all
namespaces).

Signed-off-by: p4u <pau@dabax.net>