vochain: use SignedTx proto type
Since Protobuf3 is not 100% deterministic on different implementations,
we need a new data structure model which keeps the serialized set of bytes
apart from the signature. This way we can check the signature using
exactly the same bytes sent by the Vochain clients.

This is achieved by using a new proto type SignedTx which have two
fields: TxBytes and SignatureBytes.

This commit adapts the whole vocdoni-node repository to use SignedTx. It
also declares "submitRawTx" API call public so clients can now just sent
the Raw tx to the Vochain.

Signed-off-by: p4u <pau@dabax.net>