mod: udpate go-libp2p and ipfs/go-bitswap
To include fixes for two data races that we found in go-ipfs, but which
were caused by two of its upstream libraries instead of ipfs directly.

I re-ran a smaller version of the test that was used to find both races
over a thousand times, and could no longer hit any data race.

Fixes #104.