graviton: use Value when checking a proof and some improvements
Value was ignored on the statetree implementation. Since we want to
support weighted census we need to properly use the value when checking
a proof.

In addition some minor modifications and improvements on the
gravitontree implementation (unify verify functions and add some extra
checks).

Finally, remove the eth_test.go since it slows down the testing and the
internal web3 is not used.

Signed-off-by: p4u <pau@dabax.net>