disable Tendermint tx indexer and fix scrutinizer race
Since we don't use the tendermint transaction indexer it is worth to
disable it. Its impact on performance and disk size is very important.

On scrutinizer do not execute AfetSyncBootstrap on the constructor, so
it will only be enabled if the developer wants. For testing its not
required and its the cause for some data races.

Signed-off-by: p4u <pau@dabax.net>