vochain: support costFromWeight
If costFromWeight flag is enabled, the maxCost of a ballot will be
validated against the voter weight.

Also if maxValue equals 0, the results are simplified on a single column
matrix, as the discrete approach results summary. Each row will have the
aggregated numeric results).

If maxValue=0 and costFromWeight=false, the value of the vote will
become the multiplication of the voter weight with the voted value.

Signed-off-by: p4u <pau@dabax.net>