vochain: use a better LRU
It's still safe for concurrent use, but allows us to do the common
"Get or else do something expensive and Put" without two simultaneous
opearations doing the same expensive work.

One simple way to do it would be to keep the lock grabbed while we do
the expensive computation, but that would be a pretty big bottleneck.

Instead, all LRU cache entries are wrapped by a WaitGroup, so that in
the "missing" path we can release the block immediately, while also
having concurrent Gets on the same key waiting until the value is ready.