docker: bump Go and Debian versions
And use -trimpath, since we don't want to include the GOPATH prefix of
"/go" into all paths to Go source files - it's just not useful.

While at it, remove Dockerfile.release, as nowadays the only change was
the explicit setting of CGO_ENABLED=1. It already defaults to being
enabled if GOOS and GOARCH match the host system, which is true for
docker builds.

Update the Go version in CI as well.