vochain: use addresses of fields with sync/atomic
Use field types such as int64 instead of *int64 when they are used with
sync/atomic. We can still use pointers to them, thanks to the &
operator. This helps avoid a memory indirection, which meant having to
initialise the memory explicitly, and possibly running into nil pointer
dereferences.

The & operator can never result in a nil pointer, so that's easier.
Plus, we won't need to follow a pointer which might mean a cache miss.

Fixes #178.