vochain: also use our new LRU cache for envelope heights
GetEnvelopeHeight had the same racy behavior as blockCache, since it
would do a Get, drop the mutex, do an expensive computation, and then do
an Add with a separate lock. This could result in duplicated work.

One piece of code also needs to update an existing value, rather than
just create a new one if none exists, so we update the LRU
implementation to be a bit more powerful. Now, it allows modifying an
existing value as well.