vochain: hold mempool mutex while we purge transactions
We're not certain this is the right fix, but the panic is hard to
reproduce, and this is our best guess.

We'll reopen the issue and investigate further if the panic comes up
again.

While at it, unexport the field, since we don't want others to be
calling it.

Fixes #176.