ethereum/ethevents: go back to a map+mutex
This way, the "next" method is back to being atomic. Before this change,
two concurrent calls to the method could return the same event, as we
would delete from the map separately from the ranging.

This partially reverts "ethevents: fix data race on events loop", but we
keep some changes such as the copying of the event struct in
processEventLogsFromTo, as it makes sense. We also document that copy.

Finally, improve the names a bit, to avoid using names like "e".