scrutinizer: add results cache
In addition rename the previous lru Cache to AtomicCache and
add a new type (lru.Cache) which is a wrapper of the standard
golang-lru cache.

Once results are final we store them in a LRU cache.
If not final, they will be queried to the underlaying
kv store.

In addition remove an unnecessary check on GetResults

Signed-off-by: p4u <pau@dabax.net>