censusTree: add support for multiple census
The census manager handler now receives a constructor function from the
censustree factory which allows to add and manage multiple census types.

The API exposes a new field named "censusType" which can be used by the
caller in order to choose the census type.

The default census (if not escpecified) is Graviton in order to keep
backwards compatibility.

Signed-off-by: p4u <pau@dabax.net>