ethevents: decouple EthereumHandler from EthEvents and simplify the event subscription connection logic.
EthEvents was calling NewEthereumHandler on
the object creation itself and because of this the
function usage was less flexible.
Now the EthEvents EthereumHandle is added after the
object instanciation thus making the EthEvents
subscription go-routine simpler.

The startBlock attribute was also deleted in favor
of EthereumLastKnownBlock which is an atomic value
shared among different go routines.

The former follows the subscription connection
logic simplification.
The SubscribeEthereumEventLogs function had a
weird/complex logic for reading some past events
during the initial connection and now also with
the reconnection thanks to the web3 endpoints pool,
so the new logic handles the subscription as follows:
- If subscribeOnly flag == true:
  - Get latest block and start from latestBlock - readBlocksPast
  - If latest block cannot be fetch -> Fatal
- If subscribeOnly flag = false
  - Get start block from the specs and use it as latestBlock
  - If start block does not exist for the spec -> Fatal

In both cases, once the initialization is done, the
EthereumLastKnownBlock attribute will keep track of
the latest known block of an specific web3 client and
if at some point the connection is dropped and the
subscription is initialized again, the start block for the
subscription is going to be the latest known block which is
updated every 5 seconds minus the readBlocks past for adding
some extra security.

The last change made for having the subscription connection
mechanism simplified is to share the same context with
cancel with the PrintInfo function because they share the
same web3 client. From now on the PrintInfo blocking function
is terminated if the subscription connection is not working
and reinitialized in the same goroutine that keeps track
of the ethereum events subscription.