multirpc/subpub: set limits on libp2p discovery
The library seems to generally have no default limit, and we've
encountered gateways in some environments spiking to as many as 200K
goroutines, so this seems like a likely culprit.

And even if this is not the cause, setting a limit is a good idea
anyway.

As a start, we're using four times MaxDHTpeers as the discovery limit
for both FindPeers and Advertise. It seems high enough to keep things
working, but low enough that we can't run out of memory with hundreds of
thousands of goroutines.

Updates #243.