scrutinizer: add custom big Int implementation
The JSON Marshaler for the standard type big.Int optputs an integer
number instead of an string (such as "123456789") or an hexString (such
as "0x123456). This behavior creates a problem when importing the JSON
data to a different implementation (i.e Javascript) since it forces to
read the field as an integer (instead of bigInt).

To fix this we introduce a new custom big.Int implementation which
Marshals always to a bigInt string.

Note that this fix is widely used by projects such as go-ethereum
(common/hexutil)

Signed-off-by: p4u <pau@dabax.net>