Fix block.Height appHash disparity
Don't store the block request header in the state.  Previously the block
request header was stored in the state.  I found this confusing because then
the state hash is updated at every block even if there are no transactions that
update the state, and only the height and timestamp were ever queried.

Add startBlockTimestamp in BaseApplication so that it can be consumed by the
scrutinizer (previously this was stored in the header in the state).

Store the last commited block height in the NoState of the StateDB, so that on
init we can query the height of the last stored height (required for Tendermint
Info method).  Also keep the current block in a variable in State.

Move State RWMutex into a wrapper with the Tx to make it explicit that
only the State.Tx requires locking.