Batch transactions in the StateDB
Introduce db/batch.Batch: a wrapper over a db.WriteTx that implements
db.WriteTx.  The batch acts as a WriteTx but when it becomes too big, it
automatically commits and continues with a new internal db.WriteTx
transparently.

Use this Batch instead of a regular WriteTx in the statedb.TreeTx to
avoid hitting the db.WriteTx size limit when the block contains many
updates.  Since we are now writting to disk between StateDB versions,
when opening TreeUpdates we need to make sure we start always from the
last commited version, so we query the expected root and set it via
tree.SetRoot if necessary.

This should fix the observed error "Txn is too big to fit into one
request" seen in https://github.com/vocdoni/vocdoni-node/issues/280