Store all censuses in same Database
Previously the census manager used graviton to store the censues, and
there was a different graviton database for each census.  Now we have
moved to BadgerDB with Arbo; and BadgerDB has a significant memory usage
for each opened Database.  For this reason we switch to having a single
BadgerDB database where all censuses are stored (using Arbo) under
different prefixes.  This should solve the Out of Memory issue observed
in nodes with many censues.

Introduce the concept of census version.  We define censuses previous to
the update to Arbo + BadgerDB to be version 0, and the current
implementation of censuses to be version 1.  The vocdoni-node stores
and loads censuses indexed in the namespaces.json file.  Censuses with
version 0 (note that by using json.Unmarshal, entries in the
namespaces.json that don't have the version key will default to version
= 0) are incompatible with the current implementation and will be
removed from the namespaces.json file and their corresponding database
directories will be discarded.  For safety, if entries in namespace.json
are deleted, a backup will be made with the current time timestamp, and
the corresponding database directories will be moved to a folder for
outdated censues with the current time timestamp.

Add locking on update operations in the census Trees to avoid
inconsisntencies.  Without locking, in the best case a concurrent update
would fail, and in the worst case a concurrent update would overwrite a
previous update.

For update operations in the cenuses, use db.Batch instead of db.WriteTx
to avoid encountering "txn too big" errors, which could happen with big
censues during "AddBatch" or "ImportDump" methods.  We do this by
wrapping BadgerDB via a struct that returns a db.Batch for every WriteTx
requested.