Fix height/AppHash mismatch in StateDB
Previously we were storing the new height in the StateDB at the
beginning of a block.  After commit
70d10e3f6dba1f187c1387309e5b534ba95e382f we no longer have atomic
updates for each block, so we could end with a crash in the middle of a
block, where the new height has been stored but the block has not been
completed.  After this crash, the height won't match the last state,
so the node won't be able to boot.

Fix this situation by removing the explicit storage of height in the
StateDB, and using the StateDB version as height (so we match each
commit version with the corresponding block height).  Since the StateDB
version is written last in a StateDB Commit, we should always have a
consistent state, where the stored height (StateDB version) always
matches the fully processed last block update.