db: use finalizers for sanity checks, add test APIs
Use runtime finalizers in badgerdb to catch when transactions are
garbage-collected without being discarded, and similarly when a database
is not closed properly.

That check panics if running inside a test, and otherwise just prints a
warning to stderr and does the cleanup work itself, to mitigate leaks.

Note that we don't implement these checks with pebbledb yet.

Most of the leaks found by those checks are in test code.
The only potential real leak spotted is in badgerdb's WriteTx.Commit,
which seemed to not discard the transaction if it was empty.

Also add the apis metadb.ForTest, metadb.NewTest, and
vochain.TestBaseApplication for the sake of deduplicating test logic.
This also helps all tests use the same database default and env var to
configure it. We don't use a flag as that only works on a per-package
basis.