Add db.WriteTx.Apply method
Apply applies the value-passed WriteTx into the given WriteTx,
copying the key-values from the original WriteTx into the one from
which the method is called.

For PebbleDB case, it uses the original Pebble Apply method. For the
BadgerDB case, as Badger does not offer this functionallity, our wrapper
uses reflect to access to the keys&values of the tx.

It is needed for arbo, in order to be able to work with a WriteTx from
different goroutines, where it would be done as: create a WriteTx for
each goroutine, use Apply to copy the content from the main WriteTx into
each goroutine WriteTx, proceed with the goroutines, and once completed,
use Apply to copy the content from each goroutine-WriteTx into the main
WriteTx.