add a new http router
This is an http router which reuses some code of the multirpc but
following a simpler approach. Supports not only JSON-RPC handlers but
any which implements the RouterNamespace inteface. So it will allow
creating the new API based on URL requests.

The existing router and multirpc should be replaced with this one.

Implement two routerHTTP handlers:
1. bearerstdapi for standard APIs based on URL queries and bearer authorization
2. jsonrpcapi for ECDSA signed JSON-RPC endpoints

Signed-off-by: p4u <pau@dabax.net>