testsuit: fix race condition
Due the new httpRouter implementation, the http server is available
quite before the rest of the API methods are registered into the router.

The testsuit start_test.sh script only checks if the http server is
listening or not (before starting the test). If the server is fast
enough generating the test keys (first step) it tries do execute the
addCensus API method before the gateway has it available.

This commit introduces a workaound (sleep of 10 seconds after http
server found ready) that should work for now.

Signed-off-by: p4u <pau@dabax.net>