Detect too many pre-registers in rolling census
Keep track of used weight per user in the pre-register phase.  For now
we only accept weight 1 for each RegisterKeyTx; but in theory it should
be already possible to pre-register several keys (with weight 1 each) if
the pre-register census is weighted, for example with an ERC20 census.

Resolve https://github.com/vocdoni/vocdoni-node/issues/333