enable weighted voting
Final tweaks to enable OFF_CHAIN_MERKLE_TREE weighted voting.

The census service will now return the voter value (weight)
on genProof, so the voter can use it when casting the vote.

The vochain checks the provided value if the census is Arbo
based (before a weight of 1 was hardcoded).

Extend integration tests in order to support weight.

Signed-off-by: p4u <pau@dabax.net>